% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\docType{data}
\name{census}
\alias{census}
\title{Census data}
\format{
A data frame with 1,000,000 observations on 9 variables:
\describe{
\item{LA1}{Administrative code for the city/province level.}
\item{LA2}{Administrative code for the district/county level. The city is divided into 5 districts.}
\item{LA3}{Administrative code for the sub-district/town level. In total, there are 78 sub-districts.}
\item{OA}{Census block identifier representing the smallest statistical unit. There are 2,506 census blocks in total.}
\item{gender}{Gender of the respondent.}
\item{age}{Age group in 5-year intervals.}
\item{edu}{Highest level of education attained.}
\item{mar}{Marital status.}
\item{htype}{Household type.}
}
}
\usage{
data(census)
}
\description{
\code{census} is a synthetic dataset that mimics the distribution of the
2010 Korean Census microdata. It contains 1,000,000 records, four hierarchical
administrative variables (LA1, LA2, LA3, and OA), and five key demographic
variables (gender, age, edu, mar, and htype).
}
\details{
Additional metadata for this dataset is available in the package repository:
\url{https://github.com/SLTLab-SNU/iLBA_package/blob/main/data/metadata.xlsx}
}
\keyword{datasets}
