% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_agg_freq.R
\name{get_agg_freq}
\alias{get_agg_freq}
\title{Return masked frequency for a specific cell}
\usage{
get_agg_freq(
  hkey_level,
  key,
  hkey_value,
  key_value,
  input_path = "full_tb.rds"
)
}
\arguments{
\item{hkey_level}{Integer indicating the hierarchical level to aggregate at
(e.g., \code{1} corresponds to the coarsest level).}

\item{key}{Vector of key variable names used to define
the aggregated cell.}

\item{hkey_value}{Vector of values for the hierarchical key variables used to define the aggregated cell.
Its length must match \code{hkey_level}. The values must be provided in
the same order as the hierarchical variables, from the coarsest level
to the finest level, up to the level specified by \code{hkey_level}.
For example, if \code{hkey_level = 3}, then \code{hkey_value} must contain
three values corresponding to the first three hierarchical levels.}

\item{key_value}{Vector of values for the key variables used to define the aggregated cell.
The values must be specified in the same order as the corresponding
variables in \code{key}.}

\item{input_path}{String specifying the path to the RDS object produced
by \code{save_full_tb()} (default \code{"full_tb.rds"}).}
}
\value{
Integer giving the masked count (\code{N_masked}) for the aggregated cell.
}
\description{
Applies the iLBA algorithm to a previously saved finest level frequency
table and computes the masked aggregated count for a specific cell,
identified by user-specified attribute combinations.
}
\examples{
save_full_tb(
data = census,
hkey = c("LA1","LA2","LA3","OA"),
key = c("gender", "age", "edu", "mar", "htype"),
mask_thr = 5,
output_path = "full_tb.rds"
)

get_agg_freq(
hkey_level = 3,
key = c('gender', 'age', 'edu'),
hkey_value = c('01','0104','010407'),
key_value = c(2, 4, 6),
input_path = "full_tb.rds"
)



}
