% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iLBA.R
\name{iLBA}
\alias{iLBA}
\title{Information Loss Bounded Aggregation (iLBA)}
\usage{
iLBA(x, B = 3)
}
\arguments{
\item{x}{Numeric vector of length 3: c(K, k, f_i)
\itemize{
\item K: number of original cells ≤ B
\item k: number of cells equal to B after SCA
\item f_i: true frequency from upper table
}}

\item{B}{Threshold for masking (default = 3).}
}
\value{
Numeric vector c(Masked, type1, type2).
}
\description{
Applies iLBA algorithm when aggregating cell frequencies.
This function is intended for internal use and is not meant to be called directly by users.
}
\references{
Park, M.J., Kim, H.J., & Kwon, S. (2024). \emph{Journal of the Korean Statistical Society}. Springer Nature.
}
\keyword{internal}
