% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCA.R
\name{SCA}
\alias{SCA}
\title{Small Cell Adjustment (SCA)}
\usage{
SCA(freq, B = 3)
}
\arguments{
\item{freq}{Numeric vector of cell frequencies.}

\item{B}{Threshold for adjustment (default = 3).}
}
\value{
A numeric vector of adjusted frequencies.
}
\description{
Applies a random adjustment to frequencies below a threshold B.
}
