% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\docType{data}
\name{census}
\alias{census}
\title{Census data}
\format{
A data frame with 1,000,000 observations on 9 variables:
\describe{
\item{LA1}{Administrative code for \strong{city/province}.}
\item{LA2}{Administrative code for \strong{district/county}.}
\item{LA3}{Administrative code for \strong{sub-district/town}.}
\item{OA}{Census output area code.}
\item{gender}{Gender of the respondent.}
\item{age}{Age group (categorical variable indicating 5-year or 10-year intervals).}
\item{edu}{Highest educational attainment.}
\item{mar}{Marital status.}
\item{htype}{Household type.}
}
}
\usage{
data(census)
}
\description{
Synthetic microdata derived from Statistics Korea (KOSTAT)
microdata. The dataset was generated from binned frequency
tables with small cells removed and resampled to approximate
the original distribution while ensuring anonymity. Includes
socio-demographic variables and administrative codes (LA1, LA2, LA3, OA).
}
\details{
This dataset is not the original confidential microdata.
It is an anonymized and processed version intended solely
for demonstration and examples within this package.
}
\keyword{datasets}
