% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnaggfreq.R
\name{returnaggfreq}
\alias{returnaggfreq}
\title{Return masked frequency for a specific cell}
\usage{
returnaggfreq(
  hkey.level,
  key,
  hkey.value,
  key.value,
  input.path = "fulltable.rds"
)
}
\arguments{
\item{hkey.level}{Integer indicating the number of hierarchical key levels to use (1 indicates top level).}

\item{key}{Character vector of key variable names used for the aggregated cell frequency.}

\item{hkey.value}{A vector of values for the hierarchical key variables (length must equal \code{hkey.level}).}

\item{key.value}{A vector of values for the key variables (in the same order as \code{key}).}

\item{input.path}{String path to load the RDS object produced by \code{savefulltb()} (default = "fulltable.rds").}
}
\description{
Computes the masked count via iLBA algorithm for a specified cell
defined by values of hierarchical key variables and key variables.
}
