% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_agg_tb.R
\name{save_agg_tb}
\alias{save_agg_tb}
\title{Save masked aggregated frequency table and information loss distribution}
\usage{
save_agg_tb(
  hkey_level,
  key,
  input_path = "full_tb.rds",
  output_tb_path = "agg_tb.csv",
  output_iL_path = "info_loss.csv"
)
}
\arguments{
\item{hkey_level}{Integer indicating the hierarchical level to aggregate at
(e.g., \code{1} corresponds to the coarsest level).}

\item{key}{Vector of key variable names used to define the aggregated frequency table.}

\item{input_path}{String path to the RDS produced by \code{save_full_tb()}
(default \code{"full_tb.rds"}).}

\item{output_tb_path}{String path to save the aggregated masked table
(CSV; default \code{"agg_tb.csv"}).}

\item{output_iL_path}{String path to save the information loss distribution
(CSV; default \code{"info_loss.csv"}).}
}
\value{
(Invisibly) returns a list with two elements:
\describe{
\item{agg_tb}{Aggregated masked table as a \code{data.table}.}
\item{info_loss}{Distribution of information loss as a \code{data.frame}.}
}
Also saves both objects to CSV files.
}
\description{
Applies the iLBA algorithm to a previously saved finest level frequency table to generate
a masked aggregated frequency table and the corresponding information loss distribution.
Both results are saved as CSV files.
}
\examples{
save_full_tb(
data = census,
hkey = c("LA1","LA2","LA3","OA"),
key = c("gender", "age", "edu", "mar", "htype"),
mask_thr = 5,
output_path = "full_tb.rds"
)


save_agg_tb(
hkey_level = 3,
key = c("gender","age","htype"),
input_path = "full_tb.rds",
output_tb_path = "agg_tb.csv",
output_iL_path = "info_loss.csv"
)



}
