% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savefulltb.R
\name{savefulltb}
\alias{savefulltb}
\title{Save full frequency table with true and masked cell frequencies}
\usage{
savefulltb(
  data,
  hkey,
  key = NULL,
  B = 3,
  rank = NULL,
  key.threshold = 100,
  output.path = "fulltable.rds"
)
}
\arguments{
\item{data}{Data.frame or data.table containing the raw microdata.}

\item{hkey}{Character vector of hierarchical key variable names.}

\item{key}{Character vector of key variable names (default = all column names of data except hkey).}

\item{B}{Threshold for masking (default = 3).}

\item{rank}{Numeric vector specifying the rank of hierarchical key variables (default = c(1,...,length(hkey)).}

\item{key.threshold}{Maximum number of unique values allowed for key variables (default = 100). Variables with more unique values will be removed.}

\item{output.path}{String path to save the resulting RDS object (default = "fulltable.rds").}
}
\value{
Saves a list object to an RDS file to the specified path.
The saved object contains the followings:
\item{fulltb}{A \code{data.table} with original counts (\code{N}) and masked counts (\code{N_SCA}).}
\item{B}{The masking threshold used for SCA.}
\item{hkey}{Ordered hierarchical key variable names.}
\item{hkey_values}{A named list of sorted unique values for each hierarchical key variable.}
\item{key}{Final set of key variable names used in the frequency table}
\item{key_values}{A named list of sorted unique values for each key variable.}
}
\description{
Computes a full frequency table from the microdata, applies SCA masking,
and saves the results as an RDS file.
}
