% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCA.R
\name{SCA}
\alias{SCA}
\title{Small Cell Adjustment (SCA)}
\usage{
SCA(freq, B = 3)
}
\arguments{
\item{freq}{Numeric vector of cell frequencies.}

\item{B}{Threshold for masking (default = 3).}
}
\value{
Numeric vector of masked cell frequencies.
}
\description{
Applies SCA algorithm to cell frequencies below a threshold B.
This function is intended for internal use and is not meant to be called directly by users.
}
\references{
Hundepool, A., Domingo-Ferrer, J., Franconi, L., Giessing, S., Nordholt, E. S., Spicer, K., & de Wolf, P.-P. (2012). \emph{Statistical Disclosure Control}. Wiley.
}
\keyword{internal}
